backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-celery-beat==2.9.0
dj-database-url==2.2.0
psycopg2-binary==2.9.11
celery==5.6.3
redis==7.4.0
httpx==0.27.0
anthropic>=0.95.0
openai>=1.70.0
pdfplumber==0.11.9
pytesseract==0.3.13
Pillow==11.2.1
python-docx==1.2.0
openpyxl==3.1.5
python-dotenv==1.0.1
pageindex==0.2.8
